requirements.txt
fastapi
uvicorn
GitPython
google-generativeai
pydantic
python-dotenv
pytest
httpx
